Extract viewthemeutils to commons library
