+++ app/build.gradle
@@ -332,8 +332,8 @@ dependencies {
 
     // implementation 'androidx.activity:activity-ktx:1.4.0'
 
-    //
-    implementation 'com.github.nextcloud.android-common:ui:4da3363e61ed646e091754147558ee3be346abc5'
+    // TODO substitute for tag after first common release
+    implementation 'com.github.nextcloud.android-common:ui:83c4a8aa885e8ef1bba2e44704be8625db396e1b'
 }
 
 task installGitHooks(type: Copy, group: "development") {
