+++ app/src/main/res/values-night/colors.xml
@@ -32,7 +32,6 @@
     <!-- App bar -->
     <color name="appbar">#1E1E1E</color>
     <color name="fontAppbar">#FFFFFF</color>
-    <color name="refresh_spinner_background">#222222</color>
 
     <!-- general text colors -->
     <color name="no_emphasis_text">#ffffff</color>
@@ -78,7 +77,5 @@
     <color name="dialog_background">#353535</color>
     <color name="vote_dialog_background">#424242</color>
 
-    <color name="switch_thumb_color_unchecked">#cbcbcb</color>
-    <color name="switch_track_color_unchecked">#5a5a5a</color>
 
 </resources>
