+++ settings.gradle
@@ -20,8 +20,3 @@
 
 include ':app'
 
-includeBuild('../android-common') {
-    dependencySubstitution {
-        substitute module('com.github.nextcloud.android-common:ui') using project(':ui')
-    }
-}
